refactor: group options with metadata system tables
