--- src/paimon/core/table/system/options_system_table.h
@@ -1,44 +1,0 @@
-/*
- * Copyright 2026-present Alibaba Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-
-#include <memory>
-#include <string>
-
-#include "arrow/api.h"
-#include "paimon/core/table/system/in_memory_system_table.h"
-
-namespace paimon {
-class TableSchema;
-
-/// System table for `T$options`, exposing the latest base table options as key/value rows.
-class OptionsSystemTable : public InMemorySystemTable {
- public:
-    static constexpr const char* kName = "options";
-
-    OptionsSystemTable(std::string table_path, std::shared_ptr<TableSchema> table_schema);
-
-    std::string Name() const override;
-    Result<std::shared_ptr<arrow::Schema>> ArrowSchema() const override;
-    Result<std::shared_ptr<arrow::RecordBatch>> BuildRecordBatch(
-        arrow::MemoryPool* pool) const override;
-
- private:
-    std::shared_ptr<TableSchema> table_schema_;
-};
-
-}  // namespace paimon
